+++ src/oic/utils/rp/__init__.py
@@ -11,9 +11,9 @@ from oic.oauth2 import ResponseError
 from oic.oauth2 import TokenError
 from oic.oic import AuthorizationRequest
 from oic.oic import AuthorizationResponse
-from oic.oic import ProviderConfigurationResponse
 from oic.oic import RegistrationResponse
 from oic.oic.message import OpenIDSchema
+from oic.oic.message import ProviderConfigurationResponse
 from oic.utils.authn.client import CLIENT_AUTHN_METHOD
 from oic.utils.http_util import Redirect
 from oic.utils.sanitize import sanitize