+++ src/oic/utils/rp/oauth2.py
@@ -5,14 +5,14 @@ from urllib.parse import urlsplit
 
 from oic import rndstr
 from oic.extension import client
-from oic.extension.client import ClientInfoResponse
+from oic.extension.message import ClientInfoResponse
 from oic.oauth2 import AccessTokenResponse
 from oic.oauth2 import AuthorizationRequest
 from oic.oauth2 import AuthorizationResponse
 from oic.oauth2 import ErrorResponse
 from oic.oauth2 import ResponseError
 from oic.oauth2 import TokenError
-from oic.oic import ProviderConfigurationResponse
+from oic.oic.message import ProviderConfigurationResponse
 from oic.utils.authn.client import CLIENT_AUTHN_METHOD
 from oic.utils.http_util import Redirect
 from oic.utils.sanitize import sanitize